requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
decouple==0.0.7
dj-database-url==2.3.0
Django==5.1.5
django-cloudinary-storage==0.3.0
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
